ffmpeg: Set default output format for dummy hwaccels
Dummy hwaccels, of which cuvid is the best example, behave differently
from real hwaccels. In the past, one of these behaviours was that
selecting the hwaccel would automatically cause the decoder (and
remember it's a dedicated decoder) to use the native hardware output
format.

This meant that transcoding command lines would pass frames through
device memory automatically once the right hwaccel and decoder/encoder
were selected.

With the generic hwaccel code path, dummy decoders end up following
the 'real' hwaccel path where the output format defaults to a software
format.

To avoid users facing an unexpected change in behaviour, we now
indicate whether an hwaccel is a dummy, and if it is, we set the
default output format appropriately.

To make this process easier, I updated ffmpeg to pass the HWAccel
struct to the init() call so that we know which decoder we are
dealing with.