avutil/cuda_check: Fix non-dynamic-loader implementation
The switch to an inline function means that the dynamic-loader
dependent typing can only work if the including file uses the
dynamic loader. If not, we'll get compilation errors. While not
ideal, the least complicated way to solve this is to have a
completely separate implementation when using a linked cuda
library and #ifdef against the presence of the loader.

This fixes compilation of the various SDK-linked cuda filters.