avfilter/vf_bwdif_cuda: CUDA accelerated bwdif deinterlacer
I've been sitting on this for 3 1/2 years now(!), and I finally got
around to fixing the loose ends and convincing myself that it was
correct. It follows the same basic structure as yadif_cuda, including
leaving out the edge handling, to avoid expensive branching.

One optimisation vs regular bwidf is leaving out the logic that
attempts to prefer the spatial predictor at the end. This is left over
from yadif and is not just redundant but also reduces quality. The core
bwdif algorithm already has logic for when to prefer the spatial
predictor. Thanks to cyanreg for noticing that!