avcodec/vaapi: Add support for decoding packed 8bit 4:4:4
Intel VAAPI, whether through hardware or driver decisions decodes 4:4:4
content into a packed format. In the case of 8bit content, this packed
format is not one that we support today, perhaps surprisingly.

It is described as AYUV, but empirically, it seems to be VUYA, so I
have called it that on the ffmpeg side.

Tested with mpv, decoding hevc 8bit 444 content.