lavu/pixdesc: favour formats where bit depth exactly matches
We encountered a situation with the addition of the P012 format where
the scoring of p010 and p012 when compared to yuv420p10 was identical,
leading to a situation where the ordering of which is the first tested
format and which is the second decided which is the best, even though
it's clear that p010 should be preferred as it matches the source
bit depth.

It turns out that we don't actually do any direct comparison of bit
depth as part of score calculation, which we must do if we want to tie
break this case correctly.

This change adds a small penalty of the min/max depths don't exactly
match.