hwcontext_vulkan: move VulkanFramesPriv to hwcontext_vulkan_internal
The issue is that during calls to avcodec_get_hw_frames_parameters,
avctx->internal->hwaccel_priv_data isn't guaranteed to exist.
However, we need to attach some structures to create_pnext, in order
to create frames we can decode into.
We have nowhere to put them, and they have to be valid for the entire lifetime.

Hence, we put them into the hwcontext's frames priv structure.