hwcontext_vulkan: fix map from drm for Intel VAAPI
With Intel's VAAPI implementation, it is normal to see a DRM frame
constructed as multiple layers from the same memory object. This
combination is not handled corectly right now, it tries to import the
memory once and use it twice, but as you can only specify one image
when defining the memory requirements, it doesn't import the memory
for the other images.

So, you unavoidably have to import the memory multiple times, once for
each image.

I also fixed some flag and queue family validation failures that came
up.