avfilter/vf_bwdif: Remove undesireable spatial preference logic
bwdif inherited this check from yadif, which was originally supposed to
prefer the spatial predictor if the temporal predictor was too far off.

However, the core bwdif algorithm already accounts for the spatial
predictor, so this additional check actually ends up preferring a worse
value, reducing the overall quality.

This was found by cyanreg while writing bwdif_vulkan, and the visual
improvement is pretty dramatic in some samples. If we agree that this
change is desirable, we should update all implementations.