lavu/hwcontext_vulkan: support doing queue assignment with imported vulkan device
Today, if the vulkan device is imported, all queue assignments have to
be provided as well. But if the device was created with all required
queues initialised, then it is safe for us to use the default
assignment logic, as we know any queue we pick will be available for
use.

This makes it easier for callers who don't need to write their own
assignment logic.