lavu/hwcontext: reverse order of teardown of frames context and pool
This avoids a validation failure saying an image is still in use by a
command buffer when the image is destroyed. Reversing the order
ensures the command pool and buffers are drained before images are
destroyed.

It's unclear if this is truly an error - as all the indications are
that the buffer in question is properly submitted and completed at
the time the image is freed.

This change obviously would impact all hwcontexts, but at a glance,
no context implementation seems to rely on the pool being freed before
the frames context.