avfilter/vf_nvoffruc: add support for `source_fps` expressions
The current code can only accept absolute values for the requested fps,
but it's often useful to specify a multiple of the video frame rate.

I've borrowed the code from `vf_fps` to implement this. You can read the
documentation for that filter to see how to use it, but the TLDR is you
can now use `fps=source_fps*2` or a similar expression. It also supports
the `ntsc`, `pal`, `film`, and `ntsc_film` aliases from `vf_fps`.